requirements.txt
Django==5.2.8
channels==4.1.0
daphne==4.1.2
PyMySQL==1.1.1
python-dotenv==1.0.1
google-generativeai==0.8.1
requests==2.32.3
Pillow==10.4.0
beautifulsoup4==4.12.3
openpyxl==3.1.5
reportlab==4.2.0
qrcode==8.0
python-barcode==0.15.1
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
